requirements.txt
bcrypt==4.0.1
cffi==1.15.1
contourpy==1.0.7
cryptography==39.0.1
cycler==0.11.0
dnspython==2.3.0
fonttools==4.38.0
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.0
nats-py==2.2.0
numpy==1.24.2
packaging==23.0
pandas==1.5.3
paramiko==3.0.0
Pillow==9.4.0
pycparser==2.21
pymongo==4.3.3
PyNaCl==1.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
pytz==2022.7.1
six==1.16.0
uuid==1.30
zipp==3.14.0
